Atat 61

+++ src/main/java/com/example/taskserver/controller/TaskController.java
@@ -6,12 +6,16 @@ import com.example.taskserver.domain.Task;
 import com.example.taskserver.exeption.ApiRequestExceptions;
 import com.example.taskserver.service.TaskService;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import java.util.Objects;
+
 @Slf4j
 @RestController
 @RequestMapping("/tasks")
@@ -23,7 +27,11 @@ public class TaskController {
     }
 
     @PostMapping
-    public ResponseEntity<Task> saveTask(@Valid @RequestBody Task task, Errors errors) {
+    public ResponseEntity<Task> saveTask(@Valid @RequestBody Task task, Errors errors, HttpServletRequest request) {
+        if (!service.isTokenValidBoss(request)) {
+            log.error("Not valid JWT Token");
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         log.info("Task for saving: {}", task);
         if (errors.hasErrors()) {
             log.error("Not valid task: {}", task);
@@ -38,9 +46,12 @@ public class TaskController {
     }
 
     @GetMapping
-    public ResponseEntity<?> getTask(@RequestParam(value = "taskName", required = false) String name) {
-
-        if(name==null){
+    public ResponseEntity<?> getTask(@RequestParam(value = "taskName", required = false) String name, HttpServletRequest request) {
+        if (!service.isTokenValidBoss(request)) {
+            log.error("Not valid JWT Token");
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
+        if (name == null) {
             log.info("Getting all tasks");
             return ResponseEntity.ok(service.findAllTasks());
         } else {
@@ -51,7 +62,11 @@ public class TaskController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Task> findTaskById(@PathVariable("id") Long id) {
+    public ResponseEntity<Task> findTaskById(@PathVariable("id") Long id, HttpServletRequest request) {
+        if (!service.isTokenValidBossAndUser(request)) {
+            log.error("Not valid JWT Token");
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         log.info("Searching task with id: {}", id);
         Task task = service.findById(id);
         if (Objects.isNull(task)) {
@@ -64,27 +79,44 @@ public class TaskController {
 
 
     @DeleteMapping("/{id}")
-    public void deleteTask(@PathVariable("id") Long id) {
+    public void deleteTask(@PathVariable("id") Long id ,HttpServletRequest request) {
+        if (!service.isTokenValidBoss(request)) {
+            log.error("Not valid JWT Token");
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         log.info("Deleting task with id: {}", id);
         service.delete(id);
     }
 
     @PatchMapping("/{id}")
-    public ResponseEntity<Task> updateTask(@PathVariable("id") Long id, @RequestBody Task task) {
+    public ResponseEntity<Task> updateTask(@PathVariable("id") Long id, @RequestBody Task task, HttpServletRequest request) {
+        if (!service.isTokenValidBoss(request)) {
+            log.error("Not valid JWT Token");
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         log.info("Task with id: {}. And body: {}", id, task);
         return ResponseEntity.ok(service.update(id, task));
     }
 
     @PatchMapping("/{id}/addBand")
-    public ResponseEntity<Task> addToBand(@PathVariable("id") Long id, @RequestBody Band bandName) {
+    public ResponseEntity<Task> addToBand(@PathVariable("id") Long id, @RequestBody Band bandName, HttpServletRequest request) {
+        if (!service.isTokenValidBoss(request)) {
+            log.error("Not valid JWT Token");
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         log.info("Add task with name {}", bandName);
-        service.addTaskToBand(id,bandName);
+        service.addTaskToBand(id, bandName, request);
         return ResponseEntity.ok(service.findById(id));
     }
+
     @PatchMapping("/{id}/completed")
-    public ResponseEntity<Task> makeCompleted(@PathVariable("id") Long id) {
+    public ResponseEntity<Task> makeCompleted(@PathVariable("id") Long id, HttpServletRequest request) {
+        if (!service.isTokenValidBoss(request)) {
+            log.error("Not valid JWT Token");
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         log.info("Make task with id: {} completed", id);
-        service.makeTaskCompleted(id);
+        service.makeTaskCompleted(id,request);
         return ResponseEntity.ok(service.findById(id));
     }
